+++ app/classes/web/routes/api/crafty/config/server_dir.py
@@ -107,7 +107,7 @@ class ApiCraftyConfigServerDirHandler(BaseApiHandler):
             offending_key = why.path[0] if why.path else None
             err = f"""{self.translator.translate(
                 "validators",
-                why.schema.get("error"),
+                why.schema.get("error", "additionalProperties"),
                 self.controller.users.get_user_lang_by_id(auth_data[4]["user_id"]),
             )} {offending_key}"""
             return self.finish_json(