+++ app/classes/web/routes/api/roles/role/index.py
@@ -35,6 +35,7 @@ modify_role_schema = {
             },
         },
         "manager": {"type": ["integer", "null"], "error": "roleManager"},
+        "mfa_required": {"type": "boolean", "error": "typeBool", "fill": True},
     },
     "additionalProperties": False,
     "minProperties": 1,
@@ -69,6 +70,7 @@ basic_modify_role_schema = {
                 "required": ["server_id", "permissions"],
             },
         },
+        "mfa_required": {"type": "boolean", "error": "typeBool", "fill": True},
     },
     "additionalProperties": False,
     "minProperties": 1,
@@ -201,7 +203,7 @@ class ApiRolesRoleIndexHandler(BaseApiHandler):
                 offending_key = why.path[0] if why.path else None
             err = f"""{offending_key} {self.translator.translate(
                 "validators",
-                why.schema.get("error"),
+                why.schema.get("error", "additionalProperties"),
                 self.controller.users.get_user_lang_by_id(auth_data[4]["user_id"]),
             )} {why.schema.get("enum", "")}"""
             return self.finish_json(
@@ -225,6 +227,7 @@ class ApiRolesRoleIndexHandler(BaseApiHandler):
                 data.get("name", None),
                 data.get("servers", None),
                 manager,
+                data.get("mfa_required"),
             )
         except DoesNotExist as why:
             return self.finish_json(